Make CarbonOptions keys required
Due to a new behavior of (.:?) in Data.Aeson >0.9 CarbonOptions
configuration values that are null fail to parse.

The (.:) behave in the same way in both Data.Aeson versions: it returns
Nothing if the value is null and it fails to parse if the key is not
present.

Usually configuration templates include all the keys and using (.:) seems
the appropriate choice.

See also https://github.com/bos/aeson/issues/287.